requirements.txt

blinker==1.4
certifi==2020.12.5
cffi==1.14.5
click==7.1.2
elasticsearch==7.12.0
Flask==1.1.2
Flask-Admin==1.5.8
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-SQLAlchemy==2.4.4
Flask-WTF==0.14.3
greenlet==1.0.0
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
pycparser==2.20
python-dotenv==0.17.0
six==1.15.0
SQLAlchemy==1.3.22
urllib3==1.26.4
Werkzeug==1.0.1
WTForms==2.3.3
